+++ exchanges_wrapper/exch_srv.py
@@ -52,7 +52,7 @@ def get_account(_account_name: str) -> ():
             api_auth = endpoint['api_test'] if test_net else endpoint['api_auth']
             ws_auth = endpoint['ws_test'] if test_net else endpoint['ws_auth']
             ws_public_mbr = endpoint.get('ws_public_mbr')
-            ws_api = endpoint.get('ws_api_v3_test') if test_net else endpoint.get('ws_api_v3')
+            ws_api = endpoint.get('ws_api_test') if test_net else endpoint.get('ws_api')
             #
             exchange = 'binance' if exchange == 'binance_us' else exchange
             #
@@ -203,15 +203,19 @@ class Martin(api_pb2_grpc.MartinServicer):
         response_order = api_pb2.FetchOpenOrdersResponse.Order()
         try:
             res = []
-            ws_status = bool(client.user_wss_session and client.user_wss_session.operational_status)
+            ws_status = bool(client.exchange == 'binance'
+                             and client.user_wss_session
+                             and client.user_wss_session.operational_status)
             if ws_status:
                 params = {
                     "symbol": request.symbol,
-                    "apiKey": 1,
-                    "signature": 1,
-                    "timestamp": 1
                 }
-                res = await client.user_wss_session.handle_request("openOrders.status", params)
+                res = await client.user_wss_session.handle_request(
+                    "openOrders.status",
+                    params,
+                    api_key=True,
+                    signed=True
+                )
             if not ws_status or res is None:
                 res = await client.fetch_open_orders(symbol=request.symbol, receive_window=None)
         except asyncio.CancelledError:
@@ -261,10 +265,26 @@ class Martin(api_pb2_grpc.MartinServicer):
         _queue = client.on_order_update_queues.get(request.trade_id)
         response = api_pb2.FetchOrderResponse()
         try:
-            res = await client.fetch_order(symbol=request.symbol,
-                                           order_id=request.order_id,
-                                           origin_client_order_id=None,
-                                           receive_window=None)
+            res = {}
+            ws_status = bool(client.exchange == 'binance'
+                             and client.user_wss_session
+                             and client.user_wss_session.operational_status)
+            if ws_status:
+                params = {
+                    "symbol": request.symbol,
+                    "orderId": request.order_id,
+                }
+                res = await client.user_wss_session.handle_request(
+                    "order.status",
+                    params,
+                    api_key=True,
+                    signed=True
+                )
+            if not ws_status or res is None:
+                res = await client.fetch_order(symbol=request.symbol,
+                                               order_id=request.order_id,
+                                               origin_client_order_id=None,
+                                               receive_window=None)
         except asyncio.CancelledError:
             pass  # Task cancellation should not be logged as an error
         except Exception as _ex:
@@ -300,20 +320,20 @@ class Martin(api_pb2_grpc.MartinServicer):
         response = api_pb2.SimpleResponse()
         try:
             res = []
-            ws_status = bool(client.user_wss_session and client.user_wss_session.operational_status)
-            ws_status = False
+            ws_status = bool(client.exchange == 'binance'
+                             and client.user_wss_session
+                             and client.user_wss_session.operational_status)
             if ws_status:
                 params = {
                     "symbol": request.symbol,
-                    "apiKey": 1,
-                    "signature": 1,
-                    "timestamp": 1
                 }
-                try:
-                    res = await client.user_wss_session.handle_request("openOrders.cancelAll", params)
-                except TimeoutError:
-                    ws_status = False
-            if not ws_status:
+                res = await client.user_wss_session.handle_request(
+                    "openOrders.cancelAll",
+                    params,
+                    api_key=True,
+                    signed=True
+                )
+            if not ws_status or res is None:
                 res = await client.cancel_all_orders(symbol=request.symbol, receive_window=None)
             # logger.info(f"CancelAllOrders: {res}")
         except asyncio.CancelledError:
@@ -391,7 +411,18 @@ class Martin(api_pb2_grpc.MartinServicer):
         client = open_client.client
         response = api_pb2.FetchAccountBalanceResponse()
         response_balance = api_pb2.FetchAccountBalanceResponse.Balances()
-        account_information = await client.fetch_account_information(receive_window=None)
+        account_information = {}
+        ws_status = bool(client.exchange == 'binance'
+                         and client.user_wss_session
+                         and client.user_wss_session.operational_status)
+        if ws_status:
+            account_information = await client.user_wss_session.handle_request(
+                "account.status",
+                api_key=True,
+                signed=True
+            )
+        if not ws_status or account_information is None:
+            account_information = await client.fetch_account_information(receive_window=None)
         # Send only balances
         res = account_information.get('balances', [])
         # Create consolidated list of asset balances from SPOT and Funding wallets
@@ -717,20 +748,42 @@ class Martin(api_pb2_grpc.MartinServicer):
         client = open_client.client
         # logger.info(f"CreateLimitOrder: quantity: {request.quantity}, price: {request.price}")
         try:
-            res = await client.create_order(
-                request.symbol,
-                Side.BUY if request.buy_side else Side.SELL,
-                order_type=OrderType.LIMIT,
-                time_in_force=TimeInForce.GTC,
-                quantity=request.quantity,
-                quote_order_quantity=None,
-                price=request.price,
-                new_client_order_id=request.new_client_order_id,
-                stop_price=None,
-                iceberg_quantity=None,
-                response_type=ResponseType.RESULT.value,
-                receive_window=None,
-                test=False)
+            res = {}
+            ws_status = bool(client.exchange == 'binance'
+                             and client.user_wss_session
+                             and client.user_wss_session.operational_status)
+            if ws_status:
+                params = {
+                    "symbol": request.symbol,
+                    "side": 'BUY' if request.buy_side else 'SELL',
+                    "type": 'LIMIT',
+                    "timeInForce": 'GTC',
+                    "price": request.price,
+                    "quantity": request.quantity,
+                    "newClientOrderId": request.new_client_order_id,
+                    "newOrderRespType": 'RESULT',
+                }
+                res = await client.user_wss_session.handle_request(
+                    "order.place",
+                    params,
+                    api_key=True,
+                    signed=True
+                )
+            if not ws_status or res is None:
+                res = await client.create_order(
+                    request.symbol,
+                    Side.BUY if request.buy_side else Side.SELL,
+                    order_type=OrderType.LIMIT,
+                    time_in_force=TimeInForce.GTC,
+                    quantity=request.quantity,
+                    quote_order_quantity=None,
+                    price=request.price,
+                    new_client_order_id=request.new_client_order_id,
+                    stop_price=None,
+                    iceberg_quantity=None,
+                    response_type=ResponseType.RESULT.value,
+                    receive_window=None,
+                    test=False)
         except errors.HTTPError as ex:
             logger.error(f"CreateLimitOrder for {open_client.name}:{request.symbol}:{request.new_client_order_id}"
                          f" exception: {ex}")
@@ -758,12 +811,28 @@ class Martin(api_pb2_grpc.MartinServicer):
         open_client = OpenClient.get_client(request.client_id)
         client = open_client.client
         try:
-            res = await client.cancel_order(
-                request.symbol,
-                order_id=request.order_id,
-                origin_client_order_id=None,
-                new_client_order_id=None,
-                receive_window=None)
+            res = {}
+            ws_status = bool(client.exchange == 'binance'
+                             and client.user_wss_session
+                             and client.user_wss_session.operational_status)
+            if ws_status:
+                params = {
+                    "symbol": request.symbol,
+                    "orderId": request.order_id,
+                }
+                res = await client.user_wss_session.handle_request(
+                    "order.cancel",
+                    params,
+                    api_key=True,
+                    signed=True
+                )
+            if not ws_status or res is None:
+                res = await client.cancel_order(
+                    request.symbol,
+                    order_id=request.order_id,
+                    origin_client_order_id=None,
+                    new_client_order_id=None,
+                    receive_window=None)
         except asyncio.CancelledError:
             pass  # Task cancellation should not be logged as an error
         except errors.RateLimitReached as ex: